ai-file-organizer/requirement.txt
google-genai
python-dotenv
pypdf
python-docx